Supports DPI scaling

+++ code/Cairo Desktop/CairoDesktop.Interop/NativeMethods.cs
@@ -301,7 +301,7 @@
         public struct SHFILEINFO
         {
             public IntPtr hIcon;
-            public IntPtr iIcon;
+            public int iIcon;
             public uint dwAttributes;
             [MarshalAs(UnmanagedType.ByValTStr, SizeConst = 260)]
             public string szDisplayName;
